Add gitlab ci pipeline
Add a simple pipeline with a single job to test the installation
script.

Initially it is going to run a custom centos-stream container image
, originally built and converted from centos8. Jobs to be run on
shared runners and will be migrated to specific runners later.

A few stages are defined and can be easily modified. Placeholder
scripts have been added to each of these stages to ease jobs
expansion.

The test jobs use a template so all scripts are external to the
config yml files and can be called independently.

A collateral change makes the crucible installer script executable
and calls it with ./ instead of bash.