Switch to shell executor
A specific runner has been registered with the shell executor
to run ci jobs against crucible PRs.

This change removes the container image from the gitlab config
that was used by docker executor in shared runners.

It also creates global before_script and after_script blocks to
be executed before and after each job, respectively. This is
needed to do the setup and cleanup for shell-executor runners that
reuse the same environment.