Add multibench support to all-in-one json
This patch integrates the multibench functionality into the
crucible run files, aka all-in-one json. It includes single bench
and multibench run mode support when --from-file is used. Backwards
compatibility takes care of the old run mode, which sets the
benchmark from the cli with 'crucible run <benchmark(s)> [args]'.

On multibench runs, it iterates over all mv-params objects inside
the 'benchmarks' block of the JSON file.

Although we support new --from-file and <benchmark> run modes, it
is not possible to mix both in the same command. In other words,
this is not supported: 'crucible run oslat --from-file <json>'.

A dry run mode is added through the env var CRUCIBLE_DRY_RUN=1 to
help testing debugging param construction, parsing and validation.